+++ docs/changelog.rst
@@ -4,6 +4,7 @@ Changelog
 
 Changes in 0.8.4
 ================
+- Fixed "$pull" semantics for nested ListFields (#447)
 - Allow fields to be named the same as query operators (#445)
 - Updated field filter logic - can now exclude subclass fields (#443)
 - Fixed dereference issue with embedded listfield referencefields (#439)